[GEOT-7203] Split package introduces in gt-geojson-core
