+++ modules/unsupported/geojson-core/src/main/resources/META-INF/services/org.geotools.filter.expression.PropertyAccessorFactory
@@ -1,1 +1,1 @@
-org.geotools.filter.expression.JSONNodePropertyAccessorFactory
+org.geotools.filter.expression.geojson.JSONNodePropertyAccessorFactory
