+++ modules/unsupported/geojson-core/src/test/java/org/geotools/filter/expression/geojson/JSONNodePropertyAccessorFactoryTest.java
@@ -14,7 +14,7 @@
  *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  *    Lesser General Public License for more details.
  */
-package org.geotools.filter.expression;
+package org.geotools.filter.expression.geojson;
 
 import org.junit.Assert;
 import org.junit.Before;
